Fixed "database schema has changed" error on SQLite
The error occured when using DDL in patches (which is the most common
case :)). SQLite throws that error when executing prepared statements
if another session has modified the schema since creating the statement
(in our case, mysql executes a patch that says create table). We're not
using prepared statements but the bug seems to be in PDO calling the
wrong internal SQLite functions (https://bugs.php.net/bug.php?id=43942).

Might be related to the VACUUM setting but not sure about that:
* http://www.mail-archive.com/sqlite-users@sqlite.org/msg04887.html
* http://sqlite.org/lang_vacuum.html